refactor(frontend): use correct imports
