+++ apps/frontend/app/routes/_dashboard.media.groups.$action.tsx
@@ -47,10 +47,8 @@ import {
 	DebouncedSearchInput,
 	FiltersModal,
 } from "~/components/common";
-import {
-	BaseMediaDisplayItem,
-	MetadataGroupDisplayItem,
-} from "~/components/media";
+import { BaseMediaDisplayItem } from "~/components/common";
+import { MetadataGroupDisplayItem } from "~/components/media";
 import { commaDelimitedString, pageQueryParam } from "~/lib/generals";
 import { useAppSearchParam } from "~/lib/hooks";
 import { useBulkEditCollection } from "~/lib/state/collection";
